Refactoring the page and layout for job details modal.
