--- app/(loggedin)/home/boards/[board_id]/job/[job_id]/job-info/page.tsx
@@ -1,9 +1,0 @@
-'use client';
-
-import JobInfo from '@/components/HomePage/Kanban/Column/JobPosts/JobModal/JobEdit/JobInfo';
-
-const JobDetails = () => {
-  return <JobInfo />;
-};
-
-export default JobDetails;
